use size_t instead of int

+++ src/t8_schemes/t8_default/t8_default_line/t8_dline_bits.c
@@ -357,8 +357,8 @@ t8_dline_compute_reference_coords (const t8_dline_t *elem,
                                    double *out_coords)
 {
   T8_ASSERT (t8_dline_is_valid (elem));
-  for (int coord = 0; coord < num_coords; ++coord) {
-    const int           offset = coord * skip_coords;
+  for (size_t coord = 0; coord < num_coords; ++coord) {
+    const size_t        offset = coord * skip_coords;
     out_coords[offset] = elem->x;
     out_coords[offset] += T8_DLINE_LEN (elem->level) * ref_coords[coord];
     out_coords[offset] /= (double) T8_DLINE_ROOT_LEN;
