+++ src/t8_schemes/t8_default/t8_default_vertex/t8_dvertex_bits.c
@@ -181,7 +181,7 @@ t8_dvertex_compute_reference_coords (const t8_dvertex_t *elem,
 {
   T8_ASSERT (abs (ref_coords[0]) <= T8_PRECISION_EPS);
   T8_ASSERT (t8_dvertex_is_valid (elem));
-  for (int coord = 0; coord < num_coords; ++coord) {
+  for (size_t coord = 0; coord < num_coords; ++coord) {
     out_coords[coord] = 0;
   }
 }
